#1 Add SnackBar helper function
