+++ package.json
@@ -53,11 +53,12 @@
   },
   "homepage": "https://github.com/9gag-open-source/react-native-snackbar-dialog#readme",
   "dependencies": {
-    "react-native-timer": "^1.1.3",
     "babel-cli": "^6.18.0",
+    "babel-plugin-flow-react-proptypes": "^0.19.0",
     "babel-preset-es2016": "^6.16.0",
     "babel-preset-react-native": "^1.9.1",
-    "babel-plugin-flow-react-proptypes": "^0.19.0",
+    "react-native-root-siblings": "^1.1.2",
+    "react-native-timer": "^1.1.3",
     "rimraf": "^2.5.4"
   },
   "devDependencies": {
