+++ src/index.js
@@ -1,268 +1,64 @@
-// @flow
+import React from 'react'
+import { AsyncStorage } from 'react-native'
+import RootSiblings from 'react-native-root-siblings'
 
-import React, { Component } from 'react'
-import Timer from 'react-native-timer'
+import SnackBarComponent from './SnackBar'
+import Storage from './utils/storage'
 
-import * as actions from './actions'
-import reducer from './reducer'
-import type { SnackItemType } from './type'
+export default class SnackBar {
+  static add = async (props): void => {
+    try {
+      const currentElement = await Storage.getCurrent()
 
-import {
-  View,
-  Animated,
-  Dimensions,
-  Text,
-  StyleSheet,
-  TouchableOpacity,
-  Easing,
-  InteractionManager
-} from 'react-native'
-
-const DEFAULT_DURATION: number = 5000
-const DEFAULT_FADEOUT_DURATION: number = 250
-const INITIAL_POSITION: number = -180
-
-const STYLE_BANNER_COLOR: string = '#000000'
-const TEXT_COLOR_ACCENT: string = '#0088ff'
-const HIT_SLOP = { top: 8, bottom: 8, left: 8, right: 8 }
-
-const TIMEOUT_ID: string = 'snackBar'
-
-const { width } = Dimensions.get('window')
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    position: 'absolute',
-    bottom: INITIAL_POSITION,
-    width
-  },
-
-  text: {
-    padding: 24,
-    fontSize: 16
-  },
-
-  inlineText: {
-    fontSize: 16
-  },
-
-  button: {
-    fontSize: 16,
-    fontWeight: '500'
-  },
-
-  actionRow: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-    padding: 18,
-    marginBottom: 6
-  },
-
-  inlineRow: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    padding: 18
-  },
-
-  flat: {
-    fontSize: 14,
-    marginHorizontal: 16
-  }
-})
-
-export default class SnackBar extends Component {
-  state: {
-    transformOffsetY: any,
-    transformOpacity: any
-  }
-
-  props: SnackItemType
-
-  static actions = actions
-  static reducer = reducer
-
-  static defaultProps = {
-    // Behaviour
-    fadeOutDuration: DEFAULT_FADEOUT_DURATION,
-    duration: DEFAULT_DURATION,
-    isStatic: false,
-
-    // Functions
-    onConfirm: Function,
-    onCancel: Function,
-    onAutoDismiss: Function,
-
-    // Styles
-    style: {},
-    backgroundColor: STYLE_BANNER_COLOR,
-    buttonColor: TEXT_COLOR_ACCENT,
-    textColor: 'white'
-  }
-
-  constructor (props: SnackItemType) {
-    super(props)
-
-    this.state = {
-      transformOffsetY: new Animated.Value(0),
-      transformOpacity: new Animated.Value(0)
-    }
-  }
-
-  componentDidMount () {
-    this.show()
-  }
-
-  componentWillUnmount () {
-    this.hide()
-  }
-
-  componentWillReceiveProps (nextProps: SnackItemType) {
-    const {
-      children,
-      confirmText,
-      cancelText
-    } = this.props
-
-    const isPropsChanged = (
-      children !== nextProps.children ||
-      confirmText !== nextProps.confirmText ||
-      cancelText !== nextProps.cancelText
-    )
-
-    if (isPropsChanged) {
-      this.hide()
-      this.show()
-    }
-  }
-
-  show = () => {
-    const {
-      transformOpacity,
-      transformOffsetY
-    } = this.state
-
-    const {
-      fadeOutDuration,
-      isStatic,
-      duration
-    } = this.props
-
-    Animated.parallel([
-      Animated.timing(transformOpacity, {
-        toValue: 1,
-        duration: fadeOutDuration,
-        easing: Easing.inOut(Easing.quad)
-      }),
-      Animated.timing(transformOffsetY, {
-        toValue: INITIAL_POSITION,
-        duration: fadeOutDuration,
-        easing: Easing.inOut(Easing.quad)
-      })
-    ]).start(() => {
-      if (isStatic) {
+      if (currentElement) {
+        await Storage.addToQueue(props)
         return
       }
 
-      InteractionManager.runAfterInteractions(() => {
-        Timer.setTimeout(TIMEOUT_ID, () => {
-          this.hide()
-        }, duration)
-      })
-    })
-  }
-
-  hide = () => {
-    const {
-      transformOpacity,
-      transformOffsetY
-    } = this.state
-
-    const {
-      fadeOutDuration,
-      onAutoDismiss
-    } = this.props
-
-    Animated.parallel([
-      Animated.timing(transformOpacity, {
-        toValue: 0,
-        duration: fadeOutDuration,
-        easing: Easing.inOut(Easing.quad)
-
-      }),
-      Animated.timing(transformOffsetY, {
-        toValue: INITIAL_POSITION * -1,
-        easing: Easing.inOut(Easing.quad),
-        duration: fadeOutDuration
-      })
-    ]).start(() => { onAutoDismiss && onAutoDismiss() })
-  }
-
-  renderButton = (text: string, onPress: Function = () => {}, style?: Object) => {
-    const { buttonColor } = this.props
-
-    return (
-      <TouchableOpacity hitSlop={HIT_SLOP} onPress={onPress}>
-        <Text style={[styles.button, style, { color: buttonColor }]}>
-          {text}
-        </Text>
-      </TouchableOpacity>
-    )
-  }
-
-  renderContent = () => {
-    const {
-      confirmText,
-      onConfirm,
-      cancelText,
-      onCancel,
-      children,
-      textColor
-    } = this.props
-
-    const titleElement = <Text style={[styles.text, { color: textColor }]}>{children}</Text>
-
-    if (confirmText && cancelText) {
-      return (
-        <View>
-          {titleElement}
-          <View style={styles.actionRow}>
-            { this.renderButton(cancelText, onCancel, styles.flat) }
-            { this.renderButton(confirmText, onConfirm, styles.flat) }
-          </View>
-        </View>
-      )
+      await Storage.setCurrent(new RootSiblings(<SnackBarComponent {...props} onDismiss={dismiss} />))
+    } catch (e) {
+      console.warn(e)
     }
-
-    if (confirmText) {
-      return (
-        <View style={styles.inlineRow}>
-          <Text style={[styles.inlineText, { color: textColor }]}>{children}</Text>
-          { this.renderButton(confirmText, onConfirm) }
-        </View>
-      )
-    }
-
-    return titleElement
   }
 
-  render () {
-    const { style, backgroundColor } = this.props
+  static show = async (props): void => {
+    try {
+      const newElement = <SnackBarComponent {...props} onDismiss={dismiss} />
+      const currentElement = await Storage.getCurrent()
 
-    return (
-      <Animated.View
-        style={[
-          styles.container,
-          {
-            opacity: this.state.transformOpacity,
-            transform: [{ translateY: this.state.transformOffsetY }],
-            backgroundColor
-          },
-          style
-        ]}
-      >
-        { this.renderContent() }
-      </Animated.View>
-    )
+      if (currentElement) {
+        currentElement.update(newElement)
+        return
+      }
+
+      await Storage.setCurrent(new RootSiblings(newElement))
+    } catch (e) {
+      console.warn(e)
+    }
+  }
+
+  static dismiss = async (): void => {
+    try {
+      const currentElement = await Storage.getCurrent()
+
+      if (currentElement) {
+        currentElement.destroy()
+        await Storage.removeCurrent()
+      }
+
+      const queue = await Storage.getItems()
+      const hasQueue = Array.isArray(queue) && queue.length
+
+      if (!hasQueue) {
+        return
+      }
+
+      const [currentProps, ...items] = queue
+      await Storage.setCurrent(new RootSiblings(<SnackBarComponent {...currentProps} onDismiss={dismiss} />))
+
+      items.length && await Storage.setQueue(items)
+    } catch (e) {
+      console.warn(e)
+    }
   }
 }
